pyproject.toml
[tool.poetry]
name = "ch-id-card-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Arsenii Kuznetsov <sdlmerr@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^12.2"
requests = "^2.22"

[tool.poetry.dev-dependencies]
mypy = "^0.740.0"
black = "19.3b0"
flake8 = "^3.7"
pylint = "^2.4"
isort = "^4.3"
pre-commit = "^1.20"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 120
target-version = ['py37', 'py38']
